Use the legal section in the README as the license
There is no proper license file per se in JPEG. However the README
includes the terms of distribution in a legal section call "LEGAL
ISSUES". Here we cut out that section and add it to a license file,
which we include the package. Hopefully this will help clarify the terms
of distribution.